+++ SSUT/CellDataTypeTest.cs
@@ -1,0 +1,15 @@
+﻿using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SSUT
+{
+    [TestFixture]
+    public class CellDataTypeTest : BaseTest
+    {
+
+    }
+}
